7.12: show del button only in logged-in user's blog

+++ bloglist-frontend/src/components/Blog.js
@@ -2,10 +2,10 @@ import { useState } from "react";
 import blogService from "../services/blogs";
 import { deleteBlog } from "../reducers/blogReducer";
 import { useDispatch } from "react-redux";
-// import { setNotification } from "../reducers/notificationReducer";
+import { setNotification } from "../reducers/notificationReducer";
 // import { useSelector } from "react-redux";
 
-const Blog = ({ blog, updateLikes }) => {
+const Blog = ({ blog, updateLikes, user }) => {
   const [display, setDisplay] = useState(false);
 
   //console.log(blog, "blog");
@@ -25,35 +25,36 @@ const Blog = ({ blog, updateLikes }) => {
   };
 
   const deletedBlog = async (id) => {
-    console.log(id, "id from del service");
-    await blogService.remove(id);
+    //console.log(id, "id from del service");
+    //await blogService.remove(id);
     // console.log(toRemove, "toRemove from Blog");
-    dispatch(deleteBlog(id));
+    //dispatch(deleteBlog(id));
 
-    // const del = blogs.find((blog) => blog.id === id);
+    //const del = blog.find((blog) => blog.id === id);
+
+    //setBlogs(blogs.filter((blog) => blog.id !== id));
+    // console.log(del, "itis de");
+    // console.log(blog.id, "blog id from del file");
+    const notifyMessage = window.confirm(
+      `Remove blog "${blog.title}" by ${blog.author}`
+    );
+    if (notifyMessage) {
+      await blogService.remove(id);
+      dispatch(deleteBlog(id));
+      // deleteBlog(blog.filter((blog) => blog.id !== id));
+    }
+
+    dispatch(
+      setNotification({
+        message: `"${blog.title}" blog is deleted by ${blog.author}`,
+        type: "error",
+      })
+    );
+
+    setTimeout(() => {
+      dispatch(setNotification({ message: null, type: null }));
+    }, 5000);
   };
-  //setBlogs(blogs.filter((blog) => blog.id !== id));
-  // console.log(del, "itis de");
-  // console.log(blog.id, "blog id from del file");
-  //   const notifyMessage = window.confirm(
-  //     `Remove blog "${del.title}" by ${del.author}`
-  //   );
-  //   if (notifyMessage) {
-  //     await blogService.remove(id);
-  //     setBlogs(blogs.filter((blog) => blog.id !== id));
-  //   }
-
-  //   dispatch(
-  //     setNotification({
-  //       message: `"${blog.title}" blog is deleted by ${blog.author}`,
-  //       type: "error",
-  //     })
-  //   );
-
-  //   setTimeout(() => {
-  //     dispatch(setNotification({ message: null, type: null }));
-  //   }, 5000);
-  // };
 
   return (
     <div style={blogStyle}>
@@ -79,19 +80,26 @@ const Blog = ({ blog, updateLikes }) => {
           </div>
           <div>{blog.author}</div>
           <div>
-            <button
-              id="remove-button"
-              style={{
-                color: "brown",
-                backgroundColor: "lightpink",
-                font: " bold",
-              }}
-              onClick={() => {
-                deletedBlog(blog.id);
-              }}
-            >
-              remove
-            </button>
+            {console.log(blog.user.id, "blog.user.id")}
+            {console.log(user.id, "user.id")}
+            {/* {console.log(blog.user, "blog.user")} */}
+            {/* {console.log(blog.id, "blog.id")} */}
+
+            {blog.user.id === user.id ? (
+              <button
+                id="remove-button"
+                style={{
+                  color: "brown",
+                  backgroundColor: "lightpink",
+                  font: " bold",
+                }}
+                onClick={() => {
+                  deletedBlog(blog.id);
+                }}
+              >
+                remove
+              </button>
+            ) : null}
           </div>
         </div>
       )}
@@ -115,3 +123,19 @@ export default Blog;
 //     remove
 //   </button>
 // ) : null}
+
+{
+  /* <button
+              id="remove-button"
+              style={{
+                color: "brown",
+                backgroundColor: "lightpink",
+                font: " bold",
+              }}
+              onClick={() => {
+                deletedBlog(blog.id);
+              }}
+            >
+              remove
+            </button> */
+}
